Swizzle NSApplication methods to handle exception details
