+++ AppCenterCrashes/AppCenterCrashes/Internals/Util/MSCrashesCategory.h
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MSCrashesCategory : NSObject
+
+/**
+ * Activate category for UIViewController.
+ */
++ (void)activateCategory;
+
+@end
+
+NS_ASSUME_NONNULL_END
