+++ AppCenterCrashes/AppCenterCrashes/MSCrashes.mm
@@ -12,7 +12,10 @@
 #import "MSChannelUnitConfiguration.h"
 #import "MSChannelUnitProtocol.h"
 #import "MSCrashHandlerSetupDelegate.h"
+#import "MSCrashReporter.h"
+#import "MSCrashesBufferedLog.hpp"
 #import "MSCrashesCXXExceptionWrapperException.h"
+#import "MSCrashesCategory.h"
 #import "MSCrashesDelegate.h"
 #import "MSCrashesInternal.h"
 #import "MSCrashesPrivate.h"
@@ -673,6 +676,9 @@ __attribute__((noreturn)) static void uncaught_cxx_exception_handler(const MSCra
 
     // Add a handler for C++-Exceptions.
     [MSCrashesUncaughtCXXExceptionHandlerManager addCXXExceptionHandler:uncaught_cxx_exception_handler];
+
+    // Activate category to handle additional crash details.
+    [MSCrashesCategory activateCategory];
   }
 }
