+++ AppCenterCrashes/AppCenterCrashesTests/MSErrorLogFormatterTests.mm
@@ -5,6 +5,7 @@
 
 #import "MSAppCenterInternal.h"
 #import "MSAppleErrorLog.h"
+#import "MSCrashReporter.h"
 #import "MSCrashesInternal.h"
 #import "MSCrashesPrivate.h"
 #import "MSCrashesTestUtil.h"
